+++ src/brasil/gov/portal/tests/test_init_content.py
@@ -194,36 +194,6 @@ class InitContentTestCase(unittest.TestCase):
         # Products.Doormat > 0.7
         self.assertNotIn(obj_sessao.Title(), titulos_sessoes)
 
-    def test_conteudo_link_colecao_doormat(self):
-        """Testa se funciona adicionar um link para uma coleção no rodapé."""
-        # FIXME:
-        # BBB:
-        # Quando o Products.Doormat for atualizado para 1.2, esse teste pode ser
-        # removido uma vez que ele será feito pelo próprio Doormat, em
-        # https://github.com/collective/Products.Doormat/blob/94c0e161b179c4820548a24a0804c5e65c747be5/Products/Doormat/tests/test_views.py#L49
-        # e só faz sentido porque hoje ele é customizado em brasil.gov.portal,
-        # mas a customização pode ser removida quando for atualizar essa versão.
-        # Ver
-        # https://github.com/plonegovbr/brasil.gov.portal/issues/95#issuecomment-229473089
-        ultimas_noticias_doormat = api.content.create(
-            type='DoormatCollection',
-            container=self.portal['rodape']['quarta_coluna']['navegacao'],
-            id='ultimas-noticias'
-        )
-        ultimas_noticias_doormat.setCollection(self.portal['ultimas-noticias'])
-        view = self.get_doormat_view()
-        data = view.getDoormatData()
-
-        total_collection_items = 11
-
-        total_collection_found = len([
-            link
-            for link in data[3]['column_sections'][1]['section_links']
-            if link['link_class'] == 'collection-item'
-        ])
-
-        self.assertEqual(total_collection_items, total_collection_found)
-
     def test_conteudo_doormat_removido(self):
         """Testa se o conteúdo default do Products.Doormat foi removido"""
         self.assertNotIn('doormat', self.portal.objectIds())